Reduce the firmware size and set the right shortcut for unicode
Because of the colemak setting on the OS, I have to ask the keyboard to
use ctrl+shift+i instead of ctrl+shift+u. When the keyboard sends
ctrl+shift+i it will be converted to ctrl+shift+u via OS and it will
work.

Again, it's a workaround but it works for this particular case. I hope
to find a better solution in the future